Add Multiboot documentation, kernel status summary, and shell commands
- Created MULTIBOOT.md to document Multiboot information and bootloader comparison.
- Added README-KFS1-STATUS.md summarizing the current status of KFS_1 project features.
- Introduced README.md detailing GDT and IDT concepts and their structures.
- Added gdt_flat_vs_segmented.md explaining GDT usage in flat memory models.
- Implemented shell commands in commands.c for reboot, halt, stack information, and help.
- Enhanced command processing in shell.c to handle new commands.
- Added strcmp function to utils.c for string comparison.
- Updated printf.c to support unsigned long long in print_hex.
- Removed outdated todo.md file.

+++ lib/printf.c
@@ -22,7 +22,7 @@ int	print_str(char *str)
 	return (i);
 }
 
-int	print_hex(unsigned long num, char format, int len)
+int	print_hex(unsigned long long num, char format, int len)
 {
 	char	hex_result[1024];
 	int		i;
